Add linewise general parser for question files
The general parser simply tokenizes line by line while keeping track of
the context of the lines (whether or not it's in a question). It then
delegates the work of actually storing the data to `assign()`.

Currently it will not be able to parse question paper spec files
(user parameters). In order to support this, the parsing of the
"difficulty" with its comparator must be added, in addition to
some boilerplate.